requirements.txt
streamlit
gitpython
supabase
python-dotenv
